Load TableInfo using Ion, not JSON
The result of the query `SELECT * FROM information_schema.user_tables`
is a "bag" of Ion values. In particular, each value in this bag is a
struct that has a specific schema. This commit uses that schema to
populate objects explicitly.

Before this commit, population was done by taking a roundtrip through
JSON. First, the Ion values were turned into JSON and then back into
objects. While convenient, this is not recommended as converting through
JSON is lossy. At the time of writing, this causes no problems as all
fields in the documents are types JSON supports. But there is no
guarantee that will always be the case.

As a side-effect of applying a explicit codec, we now will report any
mismatches in expected schema at the site of parsing. For example, if
had a bug where 'name' was thought to be 'tableName', this would now
fail in the `parseTableInfo` method instead of wherever the UI happened
to use the data. While the previous code *appeared* to have a typed
object, in reality the type only existed in typescript and not at
runtime.